Add files via upload
Non-invasive monitoring of drug action (NiMDA)

These scripts and data are intended to be used to build exponential models from parasite counts produced by time-lapse imaging (Fesser et al. 2020)

They intended to be used with time-lapse data from high-content imaging
The different parameters should be adapted to your needs
 Copyright (C) <2019> <Anna F. Fesser, Olivier Braissant, Francisco Olmo, John M. Kelly, Pascal Mäser, Marcel Kaiser>

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program. If not, see <https://www.gnu.org/licenses/>.